feat: add info and search

+++ src/main/java/com/mobilecourse/backend/nosql/elasticsearch/repository/EsProductRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.elasticsearch.repository.ElasticsearchRepository
  * 商品ES操作类
  * Created by macro on 2018/6/19.
  */
-public interface EsProductRepository extends ElasticsearchRepository<EsProduct, Long> {
+public interface EsProductRepository extends ElasticsearchRepository<EsProduct, Integer> {
     /**
      * 搜索查询
      *
